move federation to new infra-federation-executor module
